backend/requirements.txt
#asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
#tzdata==2023.3
djangorestframework==3.14.0
psycopg2-binary
django-rest-swagger
drf_yasg
#django-cors-headers
#black
#ddtrace
#python-dateutil
requests
